refact: favorite button을 만드는 로직을 분리합니다
